preliminary restructuring
