+++ src/io/input/FileReader.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "Particle.h"
+#include "objects/Particle.h"
 
 #include <list>
