+++ src/io/output/XYZWriter.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "Particle.h"
+#include "objects/Particle.h"
 
 #include <fstream>
 #include <list>
